@@ -1,16 +1,15 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq.AutoMock.Resolvers;
 
-namespace Moq.AutoMock.Tests.Resolvers
+namespace Moq.AutoMock.Tests.Resolvers;
+
+[TestClass]
+public class MockResolutionContextTests
 {
-    [TestClass]
-    public class MockResolutionContextTests
-    {
-        [TestMethod]
-        public void Constructor_null_checks_arguments()
-            => ConstructorTest
-            .BuildArgumentNullExceptionsTest<MockResolutionContext>()
-            .Use(new ObjectGraphContext(false))
-            .Run();
-    }
+    [TestMethod]
+    public void Constructor_null_checks_arguments()
+        => ConstructorTest
+        .BuildArgumentNullExceptionsTest<MockResolutionContext>()
+        .Use(new ObjectGraphContext(false))
+        .Run();
 }
