@@ -1,9 +1,8 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 
-namespace Moq.AutoMock.Tests.Util
+namespace Moq.AutoMock.Tests.Util;
+
+[ExcludeFromCodeCoverage]
+public class Empty
 {
-    [ExcludeFromCodeCoverage]
-    public class Empty
-    {
-    }
 }
