@@ -1,16 +1,15 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 
-namespace Moq.AutoMock.Tests.Util
+namespace Moq.AutoMock.Tests.Util;
+
+[ExcludeFromCodeCoverage]
+public class HasFuncDependencies
 {
-    [ExcludeFromCodeCoverage]
-    public class HasFuncDependencies
+    public Func<WithService> WithServiceFactory { get; }
+
+    public HasFuncDependencies(Func<WithService> withServiceFactory)
     {
-        public Func<WithService> WithServiceFactory { get; }
-        
-        public HasFuncDependencies(Func<WithService> withServiceFactory)
-        {
-            WithServiceFactory = withServiceFactory;
-        }
+        WithServiceFactory = withServiceFactory;
     }
 }
