@@ -1,7 +1,6 @@
-﻿namespace Moq.AutoMock.Tests.Util
+﻿namespace Moq.AutoMock.Tests.Util;
+
+public interface IService1
 {
-    public interface IService1
-    {
-        void Void();
-    }
+    void Void();
 }
