@@ -1,8 +1,7 @@
-﻿namespace Moq.AutoMock.Tests.Util
+﻿namespace Moq.AutoMock.Tests.Util;
+
+public interface IService2
 {
-    public interface IService2
-    {
-        IService1 Other { get; }
-        string Name { get; }
-    }
+    IService1 Other { get; }
+    string Name { get; }
 }
