@@ -1,8 +1,7 @@
-﻿namespace Moq.AutoMock.Tests.Util
-{
-    public interface IService4
-    {
-        string MainMethodName(string input);
+﻿namespace Moq.AutoMock.Tests.Util;
 
-    }
+public interface IService4
+{
+    string MainMethodName(string input);
+
 }
