@@ -1,7 +1,6 @@
-﻿namespace Moq.AutoMock.Tests.Util
+﻿namespace Moq.AutoMock.Tests.Util;
+
+public interface IService6
 {
-    public interface IService6
-    {
-        void Void(int value, string otherValue);
-    }
+    void Void(int value, string otherValue);
 }
