@@ -1,11 +1,10 @@
-﻿namespace Moq.AutoMock.Tests.Util
-{
-    public interface IService7
-    {
-        void Void(int value);
-        void Void(string value);
+﻿namespace Moq.AutoMock.Tests.Util;
 
-        object ReturnValue(int value);
-        object ReturnValue(string value);
-    }
+public interface IService7
+{
+    void Void(int value);
+    void Void(string value);
+
+    object ReturnValue(int value);
+    object ReturnValue(string value);
 }
