@@ -1,10 +1,8 @@
-﻿namespace Moq.AutoMock.Tests.Util
-{
+﻿namespace Moq.AutoMock.Tests.Util;
 
-    public interface IServiceWithPrimitives
-    {
-        long ReturnsALong();
-        long ReturnsALongWithParameter(string parameter);
-        string ReturnsAReferenceWithParameter(string parameter);
-    }
+public interface IServiceWithPrimitives
+{
+    long ReturnsALong();
+    long ReturnsALongWithParameter(string parameter);
+    string ReturnsAReferenceWithParameter(string parameter);
 }
