@@ -1,16 +1,15 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 
-namespace Moq.AutoMock.Tests.Util
-{
-    [ExcludeFromCodeCoverage]
-    public abstract class ProtectedConstructor
-    {
-        protected ProtectedConstructor() { }
-    }
+namespace Moq.AutoMock.Tests.Util;
 
-    [ExcludeFromCodeCoverage]
-    public abstract class ProtectedConstructor1
-    {
-        protected ProtectedConstructor1(IService1 _) { }
-    }
+[ExcludeFromCodeCoverage]
+public abstract class ProtectedConstructor
+{
+    protected ProtectedConstructor() { }
+}
+
+[ExcludeFromCodeCoverage]
+public abstract class ProtectedConstructor1
+{
+    protected ProtectedConstructor1(IService1 _) { }
 }
