@@ -1,15 +1,14 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 
-namespace Moq.AutoMock.Tests.Util
+namespace Moq.AutoMock.Tests.Util;
+
+[ExcludeFromCodeCoverage]
+public class WithArrayParameter
 {
-    [ExcludeFromCodeCoverage]
-    public class WithArrayParameter
-    {
-        public WithArrayParameter() { }
+    public WithArrayParameter() { }
 
 #pragma warning disable CA1801  //Parameter  is never used. Remove the parameter or use it in the method body
-        public WithArrayParameter(string[] array) { }
-        public WithArrayParameter(string[] array, string @sealed) { }
+    public WithArrayParameter(string[] array) { }
+    public WithArrayParameter(string[] array, string @sealed) { }
 #pragma warning restore CA1801
-    }
 }
