@@ -1,11 +1,10 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 
-namespace Moq.AutoMock.Tests.Util
+namespace Moq.AutoMock.Tests.Util;
+
+[ExcludeFromCodeCoverage]
+public class WithDefaultAndSingleParameter
 {
-    [ExcludeFromCodeCoverage]
-    public class WithDefaultAndSingleParameter
-    {
-        public WithDefaultAndSingleParameter() { }
-        public WithDefaultAndSingleParameter(IService1 _) { }
-    }
+    public WithDefaultAndSingleParameter() { }
+    public WithDefaultAndSingleParameter(IService1 _) { }
 }
