@@ -1,15 +1,14 @@
 using System.Diagnostics.CodeAnalysis;
 
-namespace Moq.AutoMock.Tests.Util
+namespace Moq.AutoMock.Tests.Util;
+
+[ExcludeFromCodeCoverage]
+public class WithRecursiveDependency
 {
-    [ExcludeFromCodeCoverage]
-    public class WithRecursiveDependency
+    public WithRecursiveDependency Child { get; }
+
+    public WithRecursiveDependency(WithRecursiveDependency child)
     {
-        public WithRecursiveDependency Child { get; }
-        
-        public WithRecursiveDependency(WithRecursiveDependency child)
-        {
-            Child = child;
-        }
+        Child = child;
     }
 }
