@@ -1,15 +1,14 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 
-namespace Moq.AutoMock.Tests.Util
-{
-    [ExcludeFromCodeCoverage]
-    public class WithSealedParams
-    {
-        public string Sealed { get; set; }
+namespace Moq.AutoMock.Tests.Util;
 
-        public WithSealedParams(string @sealed)
-        {
-            Sealed = @sealed;
-        }
+[ExcludeFromCodeCoverage]
+public class WithSealedParams
+{
+    public string Sealed { get; set; }
+
+    public WithSealedParams(string @sealed)
+    {
+        Sealed = @sealed;
     }
 }
