@@ -1,15 +1,14 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 
-namespace Moq.AutoMock.Tests.Util
-{
-    [ExcludeFromCodeCoverage]
-    public class WithService
-    {
-        public IService2 Service { get; set; }
+namespace Moq.AutoMock.Tests.Util;
 
-        public WithService(IService2 service)
-        {
-            Service = service;
-        }
+[ExcludeFromCodeCoverage]
+public class WithService
+{
+    public IService2 Service { get; set; }
+
+    public WithService(IService2 service)
+    {
+        Service = service;
     }
 }
