@@ -1,11 +1,9 @@
-﻿namespace Moq.AutoMock
+﻿namespace Moq.AutoMock;
+
+internal sealed class RealInstance : IInstance
 {
+    public RealInstance(object? value) => Value = value;
 
-    internal sealed class RealInstance : IInstance
-    {
-        public RealInstance(object? value) => Value = value;
-
-        public object? Value { get; }
-        public bool IsMock => false;
-    }
+    public object? Value { get; }
+    public bool IsMock => false;
 }
