@@ -1,22 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Moq.AutoMock.Resolvers
-{
-    /// <summary>
-    /// Provides the cache used by AutoMocker.
-    /// </summary>
-    public class CacheResolver : IMockResolver
-    {
-        internal Dictionary<Type, IInstance> TypeMap { get; } = new();
+namespace Moq.AutoMock.Resolvers;
 
-        /// <inheritdoc />
-        public void Resolve(MockResolutionContext context)
+/// <summary>
+/// Provides the cache used by AutoMocker.
+/// </summary>
+public class CacheResolver : IMockResolver
+{
+    internal Dictionary<Type, IInstance> TypeMap { get; } = new();
+
+    /// <inheritdoc />
+    public void Resolve(MockResolutionContext context)
+    {
+        if (TypeMap.TryGetValue(context.RequestType, out IInstance instance))
         {
-            if (TypeMap.TryGetValue(context.RequestType, out IInstance instance))
-            {
-                context.Value = instance;
-            }
+            context.Value = instance;
         }
     }
 }
