@@ -3,29 +3,28 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-namespace Moq.AutoMock.Resolvers
+namespace Moq.AutoMock.Resolvers;
+
+/// <summary>
+/// A resolver that resolves IEnumerable&lt;T&gt; requested types.
+/// </summary>
+public class EnumerableResolver : IMockResolver
 {
     /// <summary>
-    /// A resolver that resolves IEnumerable&lt;T&gt; requested types.
+    /// Resolves IEnumerable&lt;T&gt; types.
     /// </summary>
-    public class EnumerableResolver : IMockResolver
+    /// <param name="context">The resolution context.</param>
+    public void Resolve(MockResolutionContext context)
     {
-        /// <summary>
-        /// Resolves IEnumerable&lt;T&gt; types.
-        /// </summary>
-        /// <param name="context">The resolution context.</param>
-        public void Resolve(MockResolutionContext context)
-        {
-            var (am, serviceType, _) = context;
+        var (am, serviceType, _) = context;
 
-            if (!serviceType.GetTypeInfo().IsGenericType || serviceType.GetGenericTypeDefinition() != typeof(IEnumerable<>))
-                return;
+        if (!serviceType.GetTypeInfo().IsGenericType || serviceType.GetGenericTypeDefinition() != typeof(IEnumerable<>))
+            return;
 
-            var elementType = serviceType.GetGenericArguments().Single();
-            var array = Array.CreateInstance(elementType, 1);
+        var elementType = serviceType.GetGenericArguments().Single();
+        var array = Array.CreateInstance(elementType, 1);
 
-            array.SetValue(am.Get(elementType), 0);
-            context.Value = array;
-        }
+        array.SetValue(am.Get(elementType), 0);
+        context.Value = array;
     }
 }
