@@ -1,20 +1,19 @@
-﻿namespace Moq.AutoMock.Resolvers
+﻿namespace Moq.AutoMock.Resolvers;
+
+/// <summary>
+/// Resolves calls to retireve AutoMocker with itself.
+/// </summary>
+public class SelfResolver : IMockResolver
 {
     /// <summary>
-    /// Resolves calls to retireve AutoMocker with itself.
+    /// Resolves a request for AutoMocker with itself.
     /// </summary>
-    public class SelfResolver : IMockResolver
+    /// <param name="context">The mock resolution context</param>
+    public void Resolve(MockResolutionContext context)
     {
-        /// <summary>
-        /// Resolves a request for AutoMocker with itself.
-        /// </summary>
-        /// <param name="context">The mock resolution context</param>
-        public void Resolve(MockResolutionContext context)
+        if (context.RequestType == typeof(AutoMocker))
         {
-            if (context.RequestType == typeof(AutoMocker))
-            {
-                context.Value = context.AutoMocker;
-            }
+            context.Value = context.AutoMocker;
         }
     }
 }
